requirements.txt
groq==0.9.0
pydantic==1.10.15
python-dotenv==1.0.1
streamlit==1.34.0